move nodes from `graphic_element` itself to `gelement-nodes`
