+++ editor/src/messages/tool/tool_messages/artboard_tool.rs
@@ -568,7 +568,9 @@ mod test_artboard {
 			Ok(instrumented) => instrumented,
 			Err(e) => panic!("Failed to evaluate graph: {}", e),
 		};
-		instrumented.grab_all_input::<graphene_std::graphic_element::append_artboard::ArtboardInput>(&editor.runtime).collect()
+		instrumented
+			.grab_all_input::<graphene_std::element_nodes::conversion::append_artboard::ArtboardInput>(&editor.runtime)
+			.collect()
 	}
 
 	#[tokio::test]
