+++ node-graph/gelement-nodes/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod animation;
 pub mod blending_nodes;
+pub mod conversion;
 pub mod instance;
 pub mod logic;
 pub mod transform_nodes;
